borrar test que quedó por fuera del directorio backend
