+++ backend/src/test/java/app/model/notificador/NotificadorTest.java
@@ -1,47 +1,59 @@
 package app.model.notificador;
 
-import app.model.entities.*;
-import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 
+import app.model.entities.Coleccion;
+import app.model.entities.MedioComunicacion;
+import app.model.entities.MedioDeContacto;
+import app.model.entities.Perfil;
+import app.model.entities.Rol;
+import app.model.entities.Usuario;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
+import org.junit.jupiter.api.Test;
 
 class NotificadorTest {
 
-    private final Perfil perfil = new Perfil("1", null, "Test", null, List.of(), List.of());
-    private final Mensaje mensaje = new Mensaje("cuerpo", LocalDateTime.now());
+  private Perfil perfil() {
+    return new Perfil("p-1", new Usuario("u-1", Rol.USUARIO), "Lucas",
+        new Coleccion(), List.of(new MedioDeContacto(MedioComunicacion.TELEGRAM, "@lucas")),
+        new ArrayList<>());
+  }
 
-    @Test
-    void notificacion_constructor_asignaAtributos() {
-        Notificacion notificacion = new Notificacion(mensaje, perfil);
-        assertEquals(mensaje, notificacion.getMensaje());
-        assertEquals(perfil, notificacion.getUsuario());
-    }
+  @Test
+  void enviarNotificacion_delegaAlAdapter() {
+    AdapterNotificacion adapter = mock(AdapterNotificacion.class);
+    Notificador notificador = new Notificador(adapter);
+    Mensaje mensaje = new Mensaje("Hola", LocalDateTime.now());
+    Perfil perfil = perfil();
 
-    @Test
-    void notificacion_setters() {
-        Notificacion notificacion = new Notificacion(mensaje, perfil);
-        Mensaje otro = new Mensaje("otro", LocalDateTime.now());
-        notificacion.setMensaje(otro);
-        notificacion.setId("id-1");
-        assertEquals(otro, notificacion.getMensaje());
-        assertEquals("id-1", notificacion.getId());
-    }
+    notificador.enviarNotificacion(mensaje, perfil);
 
-    @Test
-    void notificador_enviarNotificacion_delegaAlAdapter() {
-        AdapterNotificacion mockAdapter = mock(AdapterNotificacion.class);
-        Notificador notificador = new Notificador(mockAdapter);
-        notificador.enviarNotificacion(mensaje, perfil);
-        verify(mockAdapter).notificar(mensaje, perfil);
-    }
+    verify(adapter).notificar(mensaje, perfil);
+  }
 
-    @Test
-    void adapterTelegram_notificar_noLanzaExcepcion() {
-        AdapterTelegram adapter = new AdapterTelegram();
-        assertDoesNotThrow(() -> adapter.notificar(mensaje, perfil));
-    }
+  @Test
+  void notificacion_guardaMensajeYUsuario() {
+    Mensaje mensaje = new Mensaje("Hola", LocalDateTime.now());
+    Perfil perfil = perfil();
+
+    Notificacion notificacion = new Notificacion(mensaje, perfil);
+
+    assertEquals(mensaje, notificacion.getMensaje());
+    assertEquals(perfil, notificacion.getUsuario());
+  }
+
+  @Test
+  void adapterTelegram_notificar_noLanzaExcepcion() {
+    AdapterTelegram adapter = new AdapterTelegram();
+    Mensaje mensaje = new Mensaje("Hola", LocalDateTime.now());
+
+    assertDoesNotThrow(
+        () -> adapter.notificar(mensaje, perfil())
+    );
+  }
 }
